refactor(handlerStrcuct):update handler structure to separete package for operations
