+++ handler/openingHandler/deleteOpening.go
@@ -1,0 +1,13 @@
+package openingHandler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func DeleteOpeningHandler(ctx *gin.Context) {
+	ctx.JSON(http.StatusOK, gin.H{
+		"message": "DeleteOpeningHandler",
+	})
+}
