+++ handler/openingHandler/listOpening.go
@@ -1,0 +1,13 @@
+package openingHandler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func ListOpeningHandler(ctx *gin.Context) {
+	ctx.JSON(http.StatusOK, gin.H{
+		"message": "ListOpeningHandler",
+	})
+}
