+++ handler/openingHandler/showOpening.go
@@ -1,0 +1,13 @@
+package openingHandler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func ShowOpeningHandler(ctx *gin.Context) {
+	ctx.JSON(http.StatusOK, gin.H{
+		"message": "ShowOpeningHandler",
+	})
+}
