+++ handler/openingHandler/updateOpening.go
@@ -1,0 +1,13 @@
+package openingHandler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func UpdateOpeningHandler(ctx *gin.Context) {
+	ctx.JSON(http.StatusOK, gin.H{
+		"message": "UpdateOpeningHandler",
+	})
+}
